very many changes

+++ app/src/main/java/com/example/vlados/crm/db/models/User.kt
@@ -5,7 +5,7 @@ import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
 
 
-data class User(val id: Int? = null,
+data class User(val id: Long? = null,
                 var shop: String? = null,
                 var username: String? = null,
                 var address: String? = null,
@@ -20,7 +20,7 @@ data class User(val id: Int? = null,
     }
 
     constructor(source: Parcel) : this(
-            source.readValue(Int::class.java.classLoader) as Int?,
+            source.readValue(Long::class.java.classLoader) as Long?,
             source.readString(),
             source.readString(),
             source.readString(),
